@@ -1,79 +1,36 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <Nullable>enable</Nullable>
         <ImplicitUsings>enable</ImplicitUsings>
-        <UserSecretsId>2c91248f-3004-4e04-88be-54a2a3bd3a75</UserSecretsId>
-        <RootNamespace>FluentCMS.Blog</RootNamespace>
+        <UserSecretsId>f83629bb-6db1-4710-aba9-0886d38e4cc6</UserSecretsId>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.6" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.8">
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-          <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
+      <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="9.0.0" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.0" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.0" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.0" />
+      <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.2" />
+      <PackageReference Include="Aspire.StackExchange.Redis.DistributedCaching" Version="9.0.0" />    
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\FluentCMS.Blog.Share\FluentCMS.Blog.Share.csproj" />
       <ProjectReference Include="..\FluentCMS\FluentCMS.csproj" />
+      <ProjectReference Include="..\FluentCMS.Blog.ServiceDefaults\FluentCMS.Blog.ServiceDefaults.csproj" />
     </ItemGroup>
     <ItemGroup>
-        <InternalsVisibleTo Include="FluentCMS.Blog.Tests" />
+      <InternalsVisibleTo Include="FluentCMS.Blog.Tests" />
     </ItemGroup>
     <ItemGroup>
-      <Content Include="..\.dockerignore">
-        <Link>.dockerignore</Link>
-      </Content>
+      <Folder Include="wwwroot\" />
     </ItemGroup>
     <ItemGroup>
       <None Update="cms.db">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
-      <None Include="wwwroot\doc\404.html" />
-      <None Include="wwwroot\doc\css\base.css" />
-      <None Include="wwwroot\doc\css\bootstrap.min.css" />
-      <None Include="wwwroot\doc\css\bootstrap.min.css.map" />
-      <None Include="wwwroot\doc\css\brands.min.css" />
-      <None Include="wwwroot\doc\css\fontawesome.min.css" />
-      <None Include="wwwroot\doc\css\solid.min.css" />
-      <None Include="wwwroot\doc\css\v4-font-face.min.css" />
-      <None Include="wwwroot\doc\img\favicon.ico" />
-      <None Include="wwwroot\doc\img\grid.png" />
-      <None Include="wwwroot\doc\index.html" />
-      <None Include="wwwroot\doc\js\base.js" />
-      <None Include="wwwroot\doc\js\bootstrap.bundle.min.js" />
-      <None Include="wwwroot\doc\js\bootstrap.bundle.min.js.map" />
-      <None Include="wwwroot\doc\js\darkmode.js" />
-      <None Include="wwwroot\doc\search\lunr.js" />
-      <None Include="wwwroot\doc\search\main.js" />
-      <None Include="wwwroot\doc\search\worker.js" />
-      <None Include="wwwroot\doc\sitemap.xml" />
-      <None Include="wwwroot\doc\sitemap.xml.gz" />
-      <None Include="wwwroot\doc\webfonts\fa-brands-400.ttf" />
-      <None Include="wwwroot\doc\webfonts\fa-brands-400.woff2" />
-      <None Include="wwwroot\doc\webfonts\fa-regular-400.ttf" />
-      <None Include="wwwroot\doc\webfonts\fa-regular-400.woff2" />
-      <None Include="wwwroot\doc\webfonts\fa-solid-900.ttf" />
-      <None Include="wwwroot\doc\webfonts\fa-solid-900.woff2" />
-      <None Include="wwwroot\doc\webfonts\fa-v4compatibility.ttf" />
-      <None Include="wwwroot\doc\webfonts\fa-v4compatibility.woff2" />
-      <None Include="wwwroot\favicon.ico" />
     </ItemGroup>
-    <ItemGroup>
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\attribute.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\constraint.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\entity.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\filter.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\filters.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\menu.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\query.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\sort.json" />
-      <_ContentIncludedByDefault Remove="wwwroot000\schema-ui\json\sorts.json" />
-    </ItemGroup>
-    <ItemGroup>
-      <Folder Include="wwwroot\files\2024-11\" />
-    </ItemGroup>
+
 </Project>
