@@ -1,4 +1,6 @@
+using FluentCMS.Test.Util;
 using FluentCMS.Utils.ApiClient;
+using FluentCMS.Utils.ResultExt;
 
 namespace FluentCMS.Blog.Tests;
 public class EntityApiTest 
@@ -22,19 +24,22 @@ public class EntityApiTest
     [Fact]
     public async Task EntityRetrieve()
     {
+        var post = Post + Guid.NewGuid().ToString("N");
+
         await _accountApiClient.EnsureLogin();
-        (await _schemaApiClient.EnsureSimpleEntity(Post, Title)).AssertSuccess();
+        (await _schemaApiClient.EnsureSimpleEntity(post, Title)).AssertSuccess();
         for (var i = 0; i < 5; i++)
         {
-            (await _entityApiClient.AddSimpleData(Post, Title, $"student{i}")).AssertSuccess();
+            (await _entityApiClient.AddSimpleData(post, Title, $"student{i}")).AssertSuccess();
         }
 
-        (await _entityApiClient.AddSimpleData(Post, Title, "good-student")).AssertSuccess();
-        (await _entityApiClient.AddSimpleData(Post, Title, "good-student")).AssertSuccess();
+        (await _entityApiClient.AddSimpleData(post, Title, "good-student")).AssertSuccess();
+        (await _entityApiClient.AddSimpleData(post, Title, "good-student")).AssertSuccess();
 
         //get first page
-        Assert.Equal(5,(await _entityApiClient.GetEntityList(Post, 0, 5)).AssertSuccess().Items.Length);
+        Assert.Equal(5,(await _entityApiClient.GetEntityList(post, 0, 5)).AssertSuccess().Items.Length);
         //get last page
-        Assert.Equal(2,(await _entityApiClient.GetEntityList(Post, 5, 5)).AssertSuccess().Items.Length);
+        var res = (await _entityApiClient.GetEntityList(post, 5, 5)).AssertSuccess();
+        Assert.Equal(2,res.Items.Length);
     }
 }
