@@ -9,11 +9,10 @@ public interface IEntitySchemaService: IEntityVectorResolver, IAttributeValueRes
 {
     Task<Result<LoadedEntity>> GetLoadedEntity(string name, CancellationToken token = default);
     Task<Entity?> GetTableDefine(string name, CancellationToken token);
-    Task<Schema> SaveTableDefine(Schema schemaDto, CancellationToken token);
-    Task<Schema> AddOrUpdateByName(Entity entity, CancellationToken token);
+    Task<Schema> SaveTableDefine(Schema schemaDto, CancellationToken ct);
+    Task<Schema> AddOrUpdateByName(Entity entity, CancellationToken ct);
     Task<Result<LoadedAttribute>> LoadOneCompoundAttribute(LoadedEntity entity, LoadedAttribute attr,HashSet<string> visitedCrosstable, CancellationToken token);
-    Task ReplaceCache();
-    bool TryGetCachedSchema(out ImmutableArray<Entity> entities);
-    Task Delete(Schema schema, CancellationToken token);
-    Task<Schema> Save(Schema schema, CancellationToken token);
+    ValueTask<ImmutableArray<Entity>> AllEntities(CancellationToken ct = default);
+    Task Delete(Schema schema, CancellationToken ct);
+    Task<Schema> Save(Schema schema, CancellationToken ct);
 }
