@@ -2,24 +2,18 @@ using Npgsql;
 
 namespace FluentCMS.Utils.DataDefinitionExecutor;
 
-public class PostgresDefinitionExecutor(string connectionString, ILogger<PostgresDefinitionExecutor> logger):IDefinitionExecutor
+public class PostgresDefinitionExecutor(NpgsqlDataSource dataSource, ILogger<PostgresDefinitionExecutor> logger):IDefinitionExecutor
 {
-    public bool TryParseDataType(string s, string type, out object? result)
+    public bool TryParseDataType(string s, string type, out DatabaseTypeValue? result)
     {
-        result = s;
-        var ret = true;
-        switch (type)
+        result = type switch
         {
-            case DataType.Int:
-                ret = int.TryParse(s, out var resultInt);
-                result = resultInt;
-                break;
-            case DataType.Datetime:
-                ret = DateTime.TryParse(s, out var resultDateTime);
-                result = resultDateTime;
-                break;
-        }
-        return ret;
+            DataType.String or DataType.Text or DataType.Na => new DatabaseTypeValue(s),
+            DataType.Int when int.TryParse(s, out var resultInt) => new DatabaseTypeValue(I: resultInt),
+            DataType.Datetime when DateTime.TryParse(s, out var resultDateTime) => new DatabaseTypeValue(D: resultDateTime),
+            _ => null
+        };
+        return result != null;
     }
 
     public async Task CreateTable(string tableName, ColumnDefinition[] columnDefinitions, CancellationToken cancellationToken)
@@ -90,25 +84,11 @@ public class PostgresDefinitionExecutor(string connectionString, ILogger<Postgre
         };
     }
 
-    private string StringToDataType(string s)
-    {
-        s = s.ToLower();
-        return s switch
-        {
-            "integer" => DataType.Int,
-            "text" => DataType.Text,
-            "timestamp" => DataType.Datetime,
-            _ => DataType.String
-        };
-    }
-    
     //use callback  instead of return QueryFactory to ensure proper disposing connection
     private async Task<T> ExecuteQuery<T>(string sql, Func<NpgsqlCommand, Task<T>> executeFunc, params (string, object)[] parameters)
     {
         logger.LogInformation(sql);
-        await using var connection = new NpgsqlConnection(connectionString);
-        await connection.OpenAsync();
-        await using var command = new NpgsqlCommand(sql, connection);
+        await using var command = dataSource.CreateCommand(sql);
 
         foreach (var (paramName, paramValue) in parameters)
         {
