@@ -2,9 +2,11 @@ using FluentResults;
 
 namespace FluentCMS.Utils.QueryBuilder;
 
+
+
 public interface IAttributeValueResolver
 {
-    public bool ResolveVal(Attribute attr, string v, out object? value);
+    public bool ResolveVal(Attribute attr, string v, out ValidValue value);
 }
 
 public interface IEntityVectorResolver
